Remove video support
Video support required native libraries which were compiled only for
armeabi and armeabi-v7a architectures, therefore the application was not
working on newer arm64 phones or any other architectures.

There was no source code provided for those libraries, so I was not able
to get the application working with video support easily and deleted the
appropriate code.

Now the application can be at least used for control of the drone via
bluetooth.